functions/requirements.txt
azure-functions
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
openai==1.12.0
python-dotenv==1.0.0
